portal-api/requirements.txt
fastapi==0.119.0
uvicorn[standard]==0.37.0
pydantic==2.12.0
python-dotenv==1.0.0
httpx==0.27.0
python-multipart==0.0.9

